FXOMDocument: add support for a custom display Node
The properties of certain JavaFX objects, such as
javafx.stage.Stage#minHeight, affects the user interface. Hence, it
would be useful if Scene Builder supports FXML files containing these
elements.

However, with the current architecture, Scene Builder can't directly
display them in the editor and allow the user to interact with them
because Scene Builder internally works by adding the root node of the
FXML file (fxomDocument.getSceneGraphRoot()) to the editor's own scene.

Fix this by teaching FxomDocument the concept of a "display node", which
is the node that will be displayed in the editor as a substitute for the
underlying root object. e.g. for javafx.scene.Scene, its "display node"
would be its root node (`scene.getRoot()`). For javafx.stage.Stage, its
"display node" would be the root node of its scene,
(`stage.getScene().getRoot()`).